# Conflicts:
#   Backend/src/main/java/com/ts/taesan/domain/transaction/api/TransactionAPI.java
#   Backend/src/main/java/com/ts/taesan/domain/transaction/repository/TransactionQRepository.java

+++ Backend/src/main/java/com/ts/taesan/global/openfeign/bank/dto/request/TransferRequest.java
@@ -18,6 +18,6 @@ public class TransferRequest {
     private String receiverAccNum;
 
     @JsonProperty("trans_amt")
-    private Integer transAmt;
+    private Long transAmt;
 
 }
